requirements.txt
aiohttp==3.6.3
asgiref==3.2.7       
astroid==2.3.3       
async-timeout==3.0.1 
attrs==20.2.0        
autopep8==1.5.6      
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2    
certifi==2020.6.20   
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
discord==1.0.1
discord.py==1.5.1
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
geographiclib==1.52
geopy==2.1.0
google-api-core==1.30.0
google-api-python-client==1.12.8
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-images-download==2.8.0
Google-Images-Search==1.3.8
googleapis-common-protos==1.53.0
googleimagedownloader==4.0.0
httplib2==0.19.1
humanfriendly==8.2
idna==2.10
instagram-private-api==1.6.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
multidict==4.7.6
numpy==1.21.0
packaging==20.9
pandas==1.3.0
Pillow==8.2.0
prettytable==0.7.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyfiglet==0.8.post1
pylint==2.4.4
PyNaCl==1.4.0
pyparsing==2.4.7
pyperclip==1.8.2
pyreadline==2.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-resize-image==1.1.19
pytz==2019.3
requests==2.24.0
requests-futures==1.0.0
requests-toolbelt==0.9.1
rsa==4.7.2
selenium==3.141.0
six==1.13.0
sockets==1.0.0
soupsieve==2.2.1
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.11.2
yarl==1.5.1
youtube-dl==2021.6.6
